Remove user store & move auth logic to plugins
